Use memory pools for wtrie node allocations.
This also greatly reduces the memory usage of the program -- it's down
to about 60% of what it used to be (presumably malloc won't allocate
small chunks, so there's waste).

However, the memory management for the trees is much more ugly now,
relying on several implicit global memory pools. It should be reworked.